Fix runtime panics in WASM bundle triggered from the account activity page (#1035)
* code improvements

* refactor account activity tabs to consume context to avoid runtime panic

* just format and lint

* fix tests

+++ src/account_activity/components.rs
@@ -181,7 +181,7 @@ pub fn AccountInternalCommandsSection(
 #[component]
 pub fn AccountDelegationsSection(
     delegations_sig: ReadSignal<Option<Vec<Option<AccountActivityQueryDelegatorExt>>>>,
-    delegator_count: Option<i64>,
+    delegator_count: Option<usize>,
     is_loading: Signal<bool>,
 ) -> impl IntoView {
     let table_columns = vec![
